# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            }
          }
        }
      }
    },
    "/projects/current": {
      "put": {
        "operationId": "put_current_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_current_project",
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/projects/current/export": {
      "post": {
        "operationId": "export_current_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Export bytes. Content-Type is `application/zip` when the format produces multiple files.",
            "content": {
              "application/octet-stream": {}
            }
          }
        }
      }
    },
    "/projects/import": {
      "post": {
        "operationId": "import_project",
        "requestBody": {
          "content": {
            "application/zip": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "delete": {
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project successfully deleted"
          },
          "400": {
            "description": "Invalid project ID"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Internal filesystem error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "ExportFormat": {
        "type": "string",
        "enum": ["khr", "psd", "rendered", "inpainted"]
      },
      "ExportProjectRequest": {
        "type": "object",
        "required": ["format"],
        "properties": {
          "defaultFont": {
            "type": ["string", "null"],
            "description": "Optional global font override (from UI preferences)."
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "pages": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "Optional subset of pages; defaults to every page."
          }
        }
      },
...
      "ListProjectsResponse": {
        "type": "object",
        "required": ["projects"],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            }
          }
        }
      },
...
      "OpenProjectRequest": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "description": "`.khrproj/` directory basename (no extension). Must exist under the\nmanaged projects directory."
          }
        }
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ProjectSummary": {
        "type": "object",
        "required": ["id", "name", "path"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable identifier — the `.khrproj` directory basename (without the\nextension). Clients address projects by this."
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Absolute filesystem path. Informational; clients never need to pass\nit back in — they use `id`."
          },
          "updatedAtMs": {
            "type": "integer",
            "format": "int64",
            "description": "Last modification time of the project directory on disk (ms since\nUNIX epoch). Used for \"recent projects\" ordering.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}